Move Jackson deserializers for bytes and uints to an infrastructure module.
