+++ data/serializer/build.gradle
@@ -5,6 +5,7 @@ dependencies {
   implementation project(':ethereum:spec')
   implementation project(':infrastructure:bls')
   implementation project(':infrastructure:bytes')
+  implementation project(':infrastructure:jackson')
   implementation 'org.apache.tuweni:tuweni-bytes'
   implementation 'org.apache.tuweni:tuweni-units'
