+++ data/serializer/src/main/java/tech/pegasys/teku/provider/JsonProvider.java
@@ -29,6 +29,22 @@ import tech.pegasys.teku.api.schema.BLSPubKey;
 import tech.pegasys.teku.api.schema.BLSSignature;
 import tech.pegasys.teku.infrastructure.bytes.Bytes20;
 import tech.pegasys.teku.infrastructure.bytes.Bytes4;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.ByteArrayDeserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.ByteArraySerializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.Bytes20Deserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.Bytes20Serializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.Bytes32Deserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.Bytes48KeyDeserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.Bytes4Deserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.Bytes4Serializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.BytesDeserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.BytesSerializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.DoubleDeserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.bytes.DoubleSerializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.uints.UInt256Deserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.uints.UInt256Serializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.uints.UInt64Deserializer;
+import tech.pegasys.teku.infrastructure.jackson.deserializers.uints.UInt64Serializer;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszBitvector;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
