+++ data/yaml/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-  implementation project(':data:serializer')
+  implementation project(':infrastructure:jackson')
   implementation 'org.apache.tuweni:tuweni-bytes'
   implementation 'com.fasterxml.jackson.core:jackson-databind'
   implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
