+++ infrastructure/jackson/build.gradle
@@ -1,0 +1,8 @@
+dependencies {
+    api 'com.fasterxml.jackson.core:jackson-databind'
+    api 'io.swagger.core.v3:swagger-annotations'
+
+    implementation project(':infrastructure:bytes')
+    implementation 'org.apache.tuweni:tuweni-bytes'
+    implementation 'org.apache.tuweni:tuweni-units'
+}
