+++ infrastructure/jackson/src/main/java/tech/pegasys/teku/infrastructure/jackson/deserializers/bytes/Bytes4Deserializer.java
@@ -11,20 +11,20 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.provider;
+package tech.pegasys.teku.infrastructure.jackson.deserializers.bytes;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import java.io.IOException;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.infrastructure.bytes.Bytes4;
 
-public class UInt64Deserializer extends JsonDeserializer<UInt64> {
+public class Bytes4Deserializer extends JsonDeserializer<Bytes4> {
 
   @Override
-  public UInt64 deserialize(JsonParser p, DeserializationContext ctxt)
+  public Bytes4 deserialize(JsonParser p, DeserializationContext ctxt)
       throws IOException, JsonProcessingException {
-    return UInt64.valueOf(p.getValueAsString());
+    return Bytes4.fromHexString(p.getValueAsString());
   }
 }
