+++ infrastructure/jackson/src/main/java/tech/pegasys/teku/infrastructure/jackson/deserializers/bytes/BytesSerializer.java
@@ -11,17 +11,17 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.provider;
+package tech.pegasys.teku.infrastructure.jackson.deserializers.bytes;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import java.io.IOException;
-import tech.pegasys.teku.infrastructure.bytes.Bytes20;
+import org.apache.tuweni.bytes.Bytes;
 
-public class Bytes20Serializer extends JsonSerializer<Bytes20> {
+public class BytesSerializer extends JsonSerializer<Bytes> {
   @Override
-  public void serialize(Bytes20 value, JsonGenerator gen, SerializerProvider serializers)
+  public void serialize(Bytes value, JsonGenerator gen, SerializerProvider provider)
       throws IOException {
     gen.writeString(value.toHexString().toLowerCase());
   }
