+++ settings.gradle
@@ -31,6 +31,7 @@ include 'infrastructure:events'
 include 'infrastructure:exceptions'
 include 'infrastructure:http'
 include 'infrastructure:io'
+include 'infrastructure:jackson'
 include 'infrastructure:json'
 include 'infrastructure:metrics'
 include 'infrastructure:unsigned'
@@ -64,3 +65,4 @@ include 'eth-tests'
 include 'eth-reference-tests'
 include 'eth-benchmark-tests'
 include 'fork-choice-tests'
+
